scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
firebase-admin==6.5.0
python-dotenv==1.0.1
deep-translator==1.11.4
Pillow>=10.3.0
lxml
